+++ packages/vtable/src/scenegraph/group-creater/cell-helper.ts
@@ -433,10 +433,11 @@ export function updateCell(col: number, row: number, table: BaseTableAPI, addNew
     colForDefine = range.start.col;
     rowForDefine = range.start.row;
   }
-  const define =
+  const define: TextColumnDefine = (
     cellLocation !== 'body'
       ? table.getHeaderDefine(colForDefine, rowForDefine)
-      : table.getBodyColumnDefine(colForDefine, rowForDefine);
+      : table.getBodyColumnDefine(colForDefine, rowForDefine)
+  ) as any;
 
   let mayHaveIcon =
     cellLocation !== 'body'
@@ -474,7 +475,12 @@ export function updateCell(col: number, row: number, table: BaseTableAPI, addNew
   cellTheme.group.cornerRadius = getCellCornerRadius(col, row, table);
 
   // fast method for text
-  if (!addNew && !isMerge && canUseFastUpdate(col, row, oldCellGroup, autoWrapText, mayHaveIcon, table)) {
+  if (
+    !addNew &&
+    !isMerge &&
+    !(define.customLayout || define.customRender || define.headerCustomLayout || define.headerCustomRender) &&
+    canUseFastUpdate(col, row, oldCellGroup, autoWrapText, mayHaveIcon, table)
+  ) {
     // update group
     const cellWidth = table.getColWidth(col);
     const cellHeight = table.getRowHeight(row);
@@ -699,8 +705,10 @@ function updateCellContent(
     // clear react container
     if (table.reactCustomLayout) {
       const reactGroup = oldCellGroup.getChildByName('custom-container');
-      const { col, row } = reactGroup;
-      table.reactCustomLayout.removeCustomCell(col, row);
+      if (reactGroup) {
+        const { col, row } = reactGroup;
+        table.reactCustomLayout.removeCustomCell(col, row);
+      }
     }
   }
   const newCellGroup = createCell(